+++ src/featureModals/Beacon/BeaconInfo.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector, shallowEqual } from 'react-redux';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 
-import { beaconClient } from './BeaconConnect';
+import { beaconClient } from './BeaconMessageRouter';
 import { setBeaconLoading } from '../../reduxContent/app/actions';
 import { createMessageAction } from '../../reduxContent/message/actions';
